requirements.txt
python-binance
pandas
numpy
ta
schedule
websocket-client